requirements.txt
matplotlib
networkx
pyyaml